+++ rocketgram/api/webhook_info.py
@@ -23,6 +23,7 @@ class WebhookInfo:
     ip_address: Optional[str]
     last_error_date: Optional[datetime]
     last_error_message: Optional[str]
+    last_synchronization_error_date: Optional[int]
     max_connections: Optional[int]
     allowed_updates: Optional[List[UpdateType]]
 
@@ -32,7 +33,10 @@ class WebhookInfo:
             return None
 
         last_error_date = datetime.utcfromtimestamp(data['last_error_date']) if 'last_error_date' in data else None
+        lsed = data['last_synchronization_error_date']
+        last_synchronization_error_date = datetime.utcfromtimestamp(lsed) if lsed in data else None
         allowed_updates = [UpdateType(m) for m in data['allowed_updates']] if 'allowed_updates' in data else None
 
         return cls(data['url'], data['has_custom_certificate'], data['pending_update_count'], data.get('ip_address'),
-                   last_error_date, data.get('last_error_message'), data.get('max_connections'), allowed_updates)
+                   last_error_date, data.get('last_error_message'), last_synchronization_error_date,
+                   data.get('max_connections'), allowed_updates)